GitHub stars data source
Data source that imports GitHub starred repositories in JSON format.

Each starred repo is imported individually, the starred repo
metadata comes from the GitHub API.

The item timestamp is set to the starred date, so they appear in the
timeline the day the repo was starred.

A small json file is saved in the timeline repository data directory
with the metadata retrieved from the GitHub API, which looks like:

```
{
 "id": 841044067,
 "name": "timelinize",
 "html_url": "https://github.com/timelinize/timelinize",
 "description": "Store your data from all your accounts and devices in a single cohesive timeline on your own computer",
 "created_at": "2024-08-11T13:27:39Z",
 "updated_at": "2024-09-03T07:17:29Z",
 "pushed_at": "2024-09-02T15:31:59Z",
 "stargazers_count": 504,
 "language": "Go",
 "full_name": "timelinize/timelinize",
 "topics": null,
 "is_template": false,
 "Topics": "archival,data-archiving,data-import,timeline",
 "private": false,
 "starred_at": "2024-08-12T17:55:48Z"
}
```

The data source currently expects the JSON file to be named like:

- ghstars.json
- ghstars-<ISO date>.json
- ghstars-<UNIX timestamp>.json