WhatsApp: Add Retrieval keys to messages
The key on the message isn't perfect, as it'll change if the person exporting their chat history has changed the name of one of the participants between exports (this would mean that participant's name would be different between exports, and their retrieval key would be different).

This seems as close as we can get without exported IDs though.

(I can't find a good way to test that the retrieval key is set properly)