Flighty: add place visits to flights
Adds Place Visits for the take off and landing airports. Includes the "Exact name" for the place, which includes terminal & gate details as well as airport name, if present.

I _haven't_ included terminal & gate info in the (default) name because the lat/long used is always for the airport (as a whole), rather than the terminal/gate.