ArchiveBox: extract PDF, Screenshot & SingleFile
Imports ArchiveBox archives. Will attach any of the PDF, the Screenshot and the SingleFile HTML snapshot (as present) to a root timestamped (new) Snapshot classification item, with a markdown description of the site which was scraped.

The attachments have a Timestamp and Timespan that exactly match the scrape start/end.

I've created the `ClassSnapshot` for this kind of "archival" item (unassociated associated with an entity), and I've uncommented the `ClassScreen` classification (the git history doesn't show why it was commented out)

This importer just ignores any archives that are aren't flagged as "status: success" or "is_archived: false" (as these are indicative of an archival in progress)