facebook: Import check-ins and tagged places; fix post attachments
We now import check-ins and tagged places, so, mainly location-related data.

Some of the concepts still need a good mapping to the project, e.g. whether to put the coordinates as the item (i.e. "I was here") or with the entity (i.e. "this place is here").

The code that imported post attachments is some of the oldest code in the base, about 10 years ago. Back then, an item could have both text and file contents, and we just combined all the various attachment data for an attachment into one related item. Now, we properly treat them as separate.

Still need to add posts in groups and FB stories.