go.mod: Upgrade mholt/archives
I guess the new Google Takeout backend uses a new/upgraded archiver that makes directories implicit within an archive. It exposed a bug in the archives package that did not return a stat entry for implicit directories. That's arguably correct, but not helpful when treating an archive as a file system or, like, a normal directory.